Card #334: support for decimal survey scores

+++ root/common/src/main/java/org/patientview/persistence/model/SurveyResponse.java
@@ -54,7 +54,7 @@ public class SurveyResponse extends BaseModel {
 
     public SurveyResponse() {}
 
-    public SurveyResponse(User user, Integer score, ScoreSeverity severity, Date date, String scoreType) {
+    public SurveyResponse(User user, Double score, ScoreSeverity severity, Date date, String scoreType) {
         this.user = user;
         this.date = date;
